бэкенд в рабочем состоянии

+++ core/src/main/java/ru/teamtwo/core/dtos/customer/CartItemDto.java
@@ -1,8 +1,8 @@
 package ru.teamtwo.core.dtos.customer;
 
 public record CartItemDto(
-    Integer id,
-    Long customerId,
-    Integer productId,
-    Integer quantity
+        Long id,
+        Long customerId,
+        Long productId,
+        Integer quantity
 ){}
