+++ core/src/main/java/ru/teamtwo/core/dtos/customer/OrderDto.java
@@ -3,7 +3,7 @@ package ru.teamtwo.core.dtos.customer;
 import java.time.Instant;
 
 public record OrderDto (
-    Integer id,
-    Integer customerId,
-    Instant createdOn
+        Long id,
+        Long customerId,
+        Instant createdOn
 ) {}
