+++ core/src/main/java/ru/teamtwo/core/dtos/customer/OrderItemDto.java
@@ -1,8 +1,8 @@
 package ru.teamtwo.core.dtos.customer;
 
 public record OrderItemDto (
-    Integer id,
-    Integer orderId,
-    Integer productOfferId,
-    Integer quantity
+        Long id,
+        Long orderId,
+        Long productOfferId,
+        Integer quantity
 ) {}
