+++ core/src/main/java/ru/teamtwo/core/dtos/product/ProductOfferDto.java
@@ -1,8 +1,8 @@
 package ru.teamtwo.core.dtos.product;
 
 public record ProductOfferDto (
-    Integer id,
-    Integer productId,
-    Integer storeId,
-    Integer quantity
+        Long id,
+        Long productId,
+        Long storeId,
+        Integer quantity
 ) {}
